pyproject.toml
[tool.poetry]
name = "Chatbot"
version = "0.0.1"
description = ""
authors = ["Coxwave"]
packages = [{include = "Chatbot"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"
openai = "^1.8.0"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
boto3 = "^1.34.29"
anthropic = "^0.25.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
black = {version = "^23.1.0", allow-prereleases = true}
ipykernel = "^6.21.1"
pytest = "7.2.0"
ruff = "^0.0.257"
pytest-asyncio = "^0.21.0"

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
